+++ docs/sampleData/lineData.js
@@ -1,0 +1,9 @@
+const lineData = [
+  { x: 0, y: 9, date: '2019/01/21 00:00:00', weekday: 'Mon' },
+  { x: 1, y: 5, date: '2019/01/22 00:00:00', weekday: 'Tue' },
+  { x: 2, y: 5, date: '2019/01/23 00:00:00', weekday: 'Wed' },
+  { x: 3, y: 3, date: '2019/01/24 00:00:00', weekday: 'Thu'},
+  { x: 4, y: 1, date: '2019/01/25 00:00:00', weekday: 'Fri' },
+];
+
+export default lineData;