+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add a simple `<LineChart>`. (#2)
+- Add `<DataLayer>` to reuse the data calculations (such as domain and range computations) logic. (#2)
+- Add `<AxisLayer>` to reuse the axes on the charts. (#2)
 - Sets up the development environment for developing React components using `Typescript`. (#1)
 - Integrates the `docz` documentation system to develop the graphics components and help developers use the library we develop. (#1)
 - Makes simple components such as `<Foo>` and `<ResponsiveLayer>` as an experiment to see if the project settings go well. (#1)